fixed branch issue with null comparison to integer

+++ test/unit_test_check_args.c
@@ -1,17 +1,17 @@
 #include "unit_test_check_args.h"
 
 int test_not_enough_args_1(void) {
-  RESULT("test_not_enough_args_1", (EXP = 0) == (ACT = check_args(0, NULL)));
+  RESULT("test_not_enough_args_1", (EXP = 0) == (ACT = (int)check_args(0, NULL)));
   return RES;
 }
 
 int test_not_enough_args_2(void) {
-  RESULT("test_not_enough_args_2", (EXP = 0) == (ACT = check_args(1, NULL)));
+  RESULT("test_not_enough_args_2", (EXP = 0) == (ACT = (int)check_args(1, NULL)));
   return RES;
 }
 
 int test_too_many_args(void) {
-  RESULT("test_too_many_args_1", (EXP = 0) == (ACT = check_args(3, NULL)));
+  RESULT("test_too_many_args_1", (EXP = 0) == (ACT = (int)check_args(3, NULL)));
   return RES;
 }
 
@@ -19,7 +19,7 @@ int test_invalid_path_1(void) {
   char s1[] = "program";
   char s2[] = "invalid.txt"; 
   char *arr[] = { s1, s2 };
-  RESULT("test_invalid_path_1", (EXP = 0) == (ACT = check_args(2, arr)));
+  RESULT("test_invalid_path_1", (EXP = 0) == (ACT = (int)check_args(2, arr)));
   return RES;
 }
 
@@ -27,7 +27,7 @@ int test_invalid_path_2(void) {
   char s1[] = "program";
   char s2[] = "text/input"; 
   char *arr[] = { s1, s2 };
-  RESULT("test_invalid_path_2", (EXP = 0) == (ACT = check_args(2, arr)));
+  RESULT("test_invalid_path_2", (EXP = 0) == (ACT = (int)check_args(2, arr)));
   return RES;
 }
 
@@ -35,14 +35,14 @@ int test_invalid_path_3(void) {
   char s1[] = "program";
   char s2[] = "mm@#!?(*\\m"; 
   char *arr[] = { s1, s2 };
-  RESULT("test_invalid_path_3", (EXP = 0) == (ACT = check_args(2, arr)));
+  RESULT("test_invalid_path_3", (EXP = 0) == (ACT = (int)check_args(2, arr)));
   return RES;
 }
 
 int test_null_path(void) {
   char s1[] = "program";
   char *arr[] = { s1, NULL };
-  RESULT("test_invalid_path_3", (EXP = 0) == (ACT = check_args(2, arr)));
+  RESULT("test_invalid_path_3", (EXP = 0) == (ACT = (int)check_args(2, arr)));
   return RES;
 }
